data type issue fix : when function return data type not align with data types of function attributes is fixed ( self aligning according to the siddhi return data types of functions)

+++ src/main/java/SiddhiAppComposites/SupportedAggregationFunctions.java
@@ -1,12 +1,12 @@
 package SiddhiAppComposites;
 
 public enum SupportedAggregationFunctions {
-    SUM("sum",SupportedDataTypes.DOUBLE),
+    SUM("sum",SupportedDataTypes.LONG),
     COUNT("count",SupportedDataTypes.LONG), // here string mean could be any type of data.
     DISTINCT_COUNT("distinctCount",SupportedDataTypes.LONG), // here String mean any type of data
     AVG("avg",SupportedDataTypes.DOUBLE),
-    MAX("max",SupportedDataTypes.DOUBLE),
-    MIN("min",SupportedDataTypes.DOUBLE),
+    MAX("max",SupportedDataTypes.LONG),
+    MIN("min",SupportedDataTypes.LONG),
     MAXFOREVER("maxForever",SupportedDataTypes.DOUBLE),
     MINFOREVER("minForever",SupportedDataTypes.DOUBLE),
     STDDEV("stdDev",SupportedDataTypes.DOUBLE);
